cleaning things up to remove warnings
